% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Title}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{a character of strings.}

\item{split}{symbol used to split.}
}
\value{
splitted_list
}
\description{
Title
}
\examples{
strsplit1("alfa,bravo,charlie,delta",split = ",")
}
